refactor publishtheme
