+++ web/src/services/theme/index.ts
@@ -5,7 +5,7 @@ export { default as metrics, metricsSizes } from "./common/metrics";
 export { default } from "./darkTheme";
 export { default as Provider } from "./provider";
 export { default as PublishedAppProvider } from "./publishedAppProvider";
-export { usePublishTheme, publishTheme, mask, type SceneThemeOptions } from "./publishTheme";
+export { usePublishTheme, mask, type SceneThemeOptions } from "./publishTheme";
 export { default as GlobalStyles } from "./common/globalStyles";
 export { default as lightTheme } from "./lightTheme";
 export { default as darkTheme } from "./darkTheme";
