+++ web/src/services/theme/publishTheme/darkTheme/colors.ts
@@ -1,0 +1,21 @@
+import { Colors } from "../types";
+
+const colors: Colors = {
+  text: {
+    strong: "#EDEDED",
+    main: "#C7C5C5",
+    weak: "#767676",
+  },
+  icon: {
+    strong: "#C6C6C6",
+    main: "#7A7777",
+    weak: "#353535",
+  },
+  other: {
+    select: "#3B3CD0",
+    mask: "#FFFFFF0D",
+    background: "#171618",
+  },
+};
+
+export default colors;
