--- web/src/services/theme/publishTheme/forest.ts
@@ -1,15 +1,0 @@
-import type { PublishTheme } from "../types";
-
-import publishColors from "./colors";
-
-export const forest: PublishTheme = {
-  strongText: publishColors.forest.text.strong,
-  mainText: publishColors.forest.text.main,
-  weakText: publishColors.forest.text.weak,
-  strongIcon: publishColors.forest.icon.strong,
-  mainIcon: publishColors.forest.icon.main,
-  weakIcon: publishColors.forest.icon.weak,
-  select: publishColors.forest.other.select,
-  mask: publishColors.forest.other.mask,
-  background: publishColors.forest.other.background,
-};
