+++ web/src/services/theme/publishTheme/forestTheme/colors.ts
@@ -1,0 +1,21 @@
+import { Colors } from "../types";
+
+const colors: Colors = {
+  text: {
+    strong: "#FFFFFF",
+    main: "#F9FEFA",
+    weak: "#66AFA2",
+  },
+  icon: {
+    strong: "#3DDCC0",
+    main: "#38B09B",
+    weak: "#1F8842",
+  },
+  other: {
+    select: "#3F886A",
+    mask: "#1F7461",
+    background: "#1B5F50",
+  },
+};
+
+export default colors;
