+++ web/src/services/theme/publishTheme/lightTheme/colors.ts
@@ -1,0 +1,21 @@
+import { Colors } from "../types";
+
+const colors: Colors = {
+  text: {
+    strong: "#FFFFFF",
+    main: "#373737",
+    weak: "#878282",
+  },
+  icon: {
+    strong: "#E9E9E9",
+    main: "#939393",
+    weak: "#C7C7C7",
+  },
+  other: {
+    select: "#F57C4B",
+    mask: "#0000001A",
+    background: "#ECECEC",
+  },
+};
+
+export default colors;
