+++ web/src/services/theme/publishTheme/lightTheme/index.ts
@@ -1,0 +1,15 @@
+import type { PublishTheme } from "../../types";
+
+import colors from "./colors";
+
+export const light: PublishTheme = {
+  strongText: colors.text.strong,
+  mainText: colors.text.main,
+  weakText: colors.text.weak,
+  strongIcon: colors.icon.strong,
+  mainIcon: colors.icon.main,
+  weakIcon: colors.icon.weak,
+  select: colors.other.select,
+  mask: colors.other.mask,
+  background: colors.other.background,
+};
